Add 'git_dep_tree'; gives java file dependencies
Script checks files that are different from master
to gather a changeset list. For each file in the
changeset list, we will look for imports and
see if we have any for other files in the same
changeset list.

We store this info in a file, per changeset file. Using
that we can report which files depend on which. The
script follows transitive dependencies (uses a sort + uniq
with a max iteration count to avoid cycles)